refactor (coin-cosmos): move HTTP calls methods to `network` folder
