+++ .changeset/good-onions-attend.md
@@ -1,0 +1,5 @@
+---
+"@ledgerhq/coin-cosmos": minor
+---
+
+refactor (coin-cosmos): move HTTP calls to `network` folder
