+++ apps/ledger-live-desktop/src/renderer/families/cosmos/AccountBalanceSummaryFooter.tsx
@@ -12,7 +12,7 @@ import Text from "~/renderer/components/Text";
 import InfoCircle from "~/renderer/icons/InfoCircle";
 import ToolTip from "~/renderer/components/Tooltip";
 import { CosmosAccount } from "@ledgerhq/live-common/families/cosmos/types";
-import { CosmosAPI } from "@ledgerhq/coin-cosmos/api/Cosmos";
+import { CosmosAPI } from "@ledgerhq/coin-cosmos/network/Cosmos";
 import { SubAccount } from "@ledgerhq/types-live";
 import { useAccountUnit } from "~/renderer/hooks/useAccountUnit";
