+++ libs/coin-modules/coin-cosmos/src/network/Cosmos.ts
@@ -15,8 +15,11 @@ import {
   CosmosRedelegation,
   CosmosTx,
   CosmosUnbonding,
+  CosmosValidatorItem,
 } from "../types";
 import * as CosmosSDKTypes from "./types";
+import { GetValidatorItem } from "./types";
+import { EnvName, EnvValue } from "@ledgerhq/live-env";
 
 const USDC_DENOM = "ibc/8E27BA2D5493AF5636760E354E46004562C46AB7EC0CC4C1CA14E9E20E2545B5";
 
@@ -32,11 +35,14 @@ export class CosmosAPI {
     return this._cosmosSDKVersion;
   }
 
-  constructor(currencyId: string) {
+  constructor(
+    currencyId: string,
+    options?: { endpoint: EnvValue<EnvName> | undefined; version: string },
+  ) {
     const crypto = cryptoFactory(currencyId);
     this.chainInstance = crypto;
-    this.defaultEndpoint = crypto.lcd;
-    this.version = crypto.version;
+    this.defaultEndpoint = options?.endpoint?.toString() ?? crypto.lcd;
+    this.version = options?.version ?? crypto.version;
   }
 
   getAccountInfo = async (
@@ -97,6 +103,29 @@ export class CosmosAPI {
     const { application_version } = await this.getNodeInfo();
     const cosmosSDKVersion = application_version.cosmos_sdk_version;
     return cosmosSDKVersion;
+  };
+
+  /**
+   * @sdk https://docs.cosmos.network/api#tag/Query/operation/Pool
+   */
+  getValidators = async (): Promise<CosmosValidatorItem[]> => {
+    const { data } = await network({
+      url: `${this.defaultEndpoint}/cosmos/staking/${this.version}/validators?status=BOND_STATUS_BONDED&pagination.limit=200`,
+      method: "GET",
+    });
+    const validators = data.validators.map((validator: GetValidatorItem) => {
+      const commission = parseFloat(validator.commission.commission_rates.rate);
+      return {
+        validatorAddress: validator.operator_address,
+        name: validator.description.moniker,
+        tokens: parseFloat(validator.tokens),
+        votingPower: 0,
+        commission,
+        estimatedYearlyRewardsRate: 0,
+      };
+    });
+
+    return validators;
   };
 
   /**
