+++ libs/coin-modules/coin-cosmos/src/network/types.ts
