+++ libs/coin-modules/coin-cosmos/src/signOperation.ts
@@ -7,7 +7,7 @@ import { encodeOperationId } from "@ledgerhq/coin-framework/operation";
 import { SignerContext } from "@ledgerhq/coin-framework/signer";
 import { ExpertModeRequired, UserRefusedOnDevice } from "@ledgerhq/errors";
 import type { AccountBridge, Operation, OperationType } from "@ledgerhq/types-live";
-import { CosmosAPI } from "./api/Cosmos";
+import { CosmosAPI } from "./network/Cosmos";
 import { buildTransaction, txToMessages } from "./buildTransaction";
 import cryptoFactory from "./chain/chain";
 import { CosmosAccount, RETURN_CODES, Transaction } from "./types";
