+++ libs/coin-modules/coin-cosmos/src/synchronisation.unit.test.ts
@@ -3,11 +3,11 @@ import * as jsHelpers from "@ledgerhq/coin-framework/bridge/jsHelpers";
 import { AccountShapeInfo } from "@ledgerhq/coin-framework/bridge/jsHelpers";
 import { CryptoCurrency } from "@ledgerhq/types-cryptoassets";
 import { Operation, SyncConfig } from "@ledgerhq/types-live";
-import { CosmosAPI } from "./api/Cosmos";
+import { CosmosAPI } from "./network/Cosmos";
 import { getAccountShape } from "./synchronisation";
 import { CosmosAccount, CosmosOperation, CosmosTx } from "./types";
 
-jest.mock("./api/Cosmos");
+jest.mock("./network/Cosmos");
 jest.mock("@ledgerhq/coin-framework/account");
 jest.mock("@ledgerhq/coin-framework/bridge/jsHelpers");
